(PC-21183) refactor(utils): move all date utils from bookOffer to shared folder
