+++ src/shared/date/getPastYears.ts
@@ -1,0 +1,7 @@
+import range from 'lodash/range'
+
+export function getPastYears(startYear: number, currentYear: string) {
+  return range(startYear, parseInt(currentYear) + 1, 1)
+    .reverse()
+    .map((year) => year.toString())
+}
