+++ src/shared/date/months.ts
@@ -1,0 +1,29 @@
+export const monthNames = [
+  'Janvier',
+  'Février',
+  'Mars',
+  'Avril',
+  'Mai',
+  'Juin',
+  'Juillet',
+  'Août',
+  'Septembre',
+  'Octobre',
+  'Novembre',
+  'Décembre',
+]
+
+export const monthNamesShort = [
+  'Janv.',
+  'Févr.',
+  'Mars',
+  'Avril',
+  'Mai',
+  'Juin',
+  'Juil.',
+  'Août',
+  'Sept.',
+  'Oct.',
+  'Nov.',
+  'Déc.',
+]
