+++ src/ui/components/inputs/DropDown/DropDown.native.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { monthNames } from 'features/bookOffer/components/Calendar/Calendar.utils'
+import { monthNames } from 'shared/date/months'
 import { render } from 'tests/utils'
 import { DropDown } from 'ui/components/inputs/DropDown/DropDown'
